pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "spanishgas"
version = "0.1.0"
description = "SpanishGas decision intelligence platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.1",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "pyarrow>=14.0",
    "boto3>=1.28",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "streamlit>=1.28",
    "plotly>=5.18",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "moto[s3,dynamodb,stepfunctions,sagemaker]>=5.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.setuptools.packages.find]
include = ["src*", "configs*"]
